Improvment of code structure
* Add json naming in group model
* Move rest handler to rest a rest package
