+++ .cfignore
@@ -1,0 +1,12 @@
+cmd
+api
+models
+playground
+vendor
+app.yml
+app.json
+docker-compose.yml
+Dockerfile
+Gopkg.lock
+Gopkg.toml
+README.md
