+++ api/controllers/rest/group_handler.go
@@ -1,4 +1,4 @@
-package controllers
+package rest
 
 import (
 	"github.com/labstack/echo"
