+++ api/dao/mariadb/user_mysql.go
@@ -20,10 +20,8 @@ func NewMysqlUserRepository(db *gorm.DB) repo.UserRepository {
 }
 
 func (r *userRepository) FindAll() ([]*models.User, error) {
-
 	var users []*models.User
 	r.db.Find(&users)
-
 	return users, r.db.Error
 }
 
@@ -32,9 +30,7 @@ func (r *userRepository) FindOne(id string) (*models.User, error) {
 
 	// convert string to int
 	var userId, _ = strconv.Atoi(id)
-
 	var users []*models.User
-
 	r.db.First(&users, userId)
 
 	if len(users) > 0 {
@@ -50,17 +46,12 @@ func (r *userRepository) Save(user *models.User) *models.User {
 }
 
 func (r *userRepository) Update(user *models.User) (*models.User, error) {
-
 	r.db.Update(&user)
-
 	return user, r.db.Error
 }
 
 func (r *userRepository) Delete(id int) error {
-
 	log.WithFields(log.Fields{"user_id": id}).Debug("delete user from mysql db")
-
 	r.db.Where("id = ?", id).Delete(&models.User{})
-
 	return r.db.Error
 }
