+++ cmd/biergit/main.go
@@ -19,19 +19,15 @@ import (
 	"context"
 	"flag"
 	"git.skydevelopment.ch/zrh-dev/go-basics/api/controllers"
-	"git.skydevelopment.ch/zrh-dev/go-basics/api/dao/mariadb"
-	"git.skydevelopment.ch/zrh-dev/go-basics/api/dao/mongodb"
+	"git.skydevelopment.ch/zrh-dev/go-basics/api/dao"
 	"git.skydevelopment.ch/zrh-dev/go-basics/api/repo"
 	"git.skydevelopment.ch/zrh-dev/go-basics/api/services"
 	"git.skydevelopment.ch/zrh-dev/go-basics/config"
 	"git.skydevelopment.ch/zrh-dev/go-basics/models"
 	"git.skydevelopment.ch/zrh-dev/go-basics/playground"
 	_ "github.com/go-sql-driver/mysql"
-	"github.com/jinzhu/gorm"
-	"github.com/mongodb/mongo-go-driver/mongo"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
-	"net/url"
 	"os"
 	"os/signal"
 	"strconv"
@@ -44,7 +40,7 @@ var persons = []string{"jan", "test1", "test2"}
 var conf *viper.Viper
 
 // Datasource Selection (mariadb, mongodb)
-const database = "mariadb"
+const Database = "mariadb"
 
 func main() {
 
@@ -68,11 +64,11 @@ func main() {
 	var transactionRepo repo.TransactionRepository
 
 	// Check which database should be used and initiate it
-	switch database {
+	switch Database {
 	case "mongodb":
-		userRepo = establishMongoDbConnection()
+		userRepo = dao.EstablishMongoDbConnection(conf)
 	case "mariadb":
-		userRepo, groupRepo, transactionRepo = establishMariaDbConnection()
+		userRepo, groupRepo, transactionRepo = dao.EstablishMariaDbConnection(conf)
 	}
 
 	// Create Service
@@ -81,7 +77,7 @@ func main() {
 	transactionService := services.NewTransactionService(transactionRepo)
 
 	// Add Mock Data
-	addMockData(userService, groupService, transactionService)
+	dao.AddMockData(userService, groupService, transactionService)
 
 	// Create HTTP Server
 	httpServer := controllers.NewServer(userService, groupService, transactionService, conf)
@@ -113,115 +109,6 @@ func main() {
 
 	log.Println("shutting down")
 	os.Exit(0)
-}
-
-
-func addMockData(userService services.UserService, groupService services.GroupService, transactionService services.TransactionService) {
-
-	log.Debug("Add Mock data")
-
-	// Add some Mock Data
-	jan := models.User{
-		FirstName: "Jan",
-		LastName:  "Minder",
-	}
-
-	// Add some Mock Data
-	luca := models.User{
-		FirstName: "Luca",
-		LastName:  "Hostetter",
-	}
-
-	// Add some Mock Data
-	biergit := models.Group{
-		GroupName: "biergit",
-	}
-
-	// Add some Mock Data
-	bspTrans := models.Transaction{
-		Source: jan,
-		Target:  luca,
-		Context: biergit,
-		Amount: 2,
-	}
-
-	log.Debug(bspTrans)
-
-	// Create Services
-	userService.CreateUser(&jan)
-	groupService.CreateGroup(&biergit)
-	transactionService.CreateTransaction(&bspTrans)
-}
-
-func establishMongoDbConnection() repo.UserRepository {
-	// Start Mongo DB Connection
-	client, err := mongo.Connect(context.TODO(), "mongodb://" + conf.GetString("mongodb.host") + ":" + strconv.Itoa(conf.GetInt("mongodb.port")))
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Check the connection
-	err = client.Ping(context.TODO(), nil)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	userCollection := client.Database("biergit").Collection("users")
-	userRepo := mongodb.NewMongoDbUserRepository(userCollection)
-
-	log.Debug("Connected to MongoDB!")
-
-	return userRepo
-}
-
-// Establish a new connection to the configured mariaDB and initiate all repositories
-func establishMariaDbConnection() (repo.UserRepository, repo.GroupRepository, repo.TransactionRepository)  {
-
-	var db *gorm.DB
-	var err error
-
-	// Check if uri is configured
-	dsn := conf.GetString("mariadb.jdbc")
-
-	log.Debug("DSN: ", dsn)
-
-	if dsn != "" {
-		// Parse the DSN String
-		dbUrl, _ := url.Parse(dsn)
-		log.Debug("QueryString: ", dbUrl.RawQuery)
-
-		// Create new MYSQL Connection
-		db, err = gorm.Open(dbUrl.Scheme, dbUrl.User.String() + "@tcp(" +
-			dbUrl.Host + ")" +
-			dbUrl.Path + "?")
-		// TODO: Check unknown system variable of reconnect flag
-
-	} else {
-		// Create new MYSQL Connection based on configuration file
-		db, err = gorm.Open("mysql", conf.GetString("mariadb.user") + ":" +
-			conf.GetString("mariadb.password") + "@tcp(" +
-			conf.GetString("mariadb.host") + ":" +
-			strconv.Itoa(conf.GetInt("mariadb.port")) + ")/" +
-			conf.GetString("mariadb.schema") + "?charset=utf8&parseTime=True")
-	}
-
-	if err != nil {
-		log.Fatal("Failed to connect to database: ", err)
-	}
-
-	// Enable SQL Query Logs
-	db.LogMode(true)
-
-	// Migrate Database
-	migrateRelationalDB(db)
-
-	userRepo := mariadb.NewMysqlUserRepository(db)
-	groupRepo := mariadb.NewMysqlGroupRepository(db)
-	transactionRepo := mariadb.NewMysqlTransactionRepository(db)
-
-	return userRepo, groupRepo, transactionRepo
 }
 
 // Test Basic GO Principles
@@ -268,12 +155,4 @@ func basicPrinciples() {
 func setupLogger() {
 	log.SetFormatter(&log.TextFormatter {})
 	log.SetLevel(log.DebugLevel)
-}
-
-// Migrate all Releational Database Tables
-func migrateRelationalDB(db *gorm.DB) {
-	log.Debug("Migrating Database Schema")
-	db.AutoMigrate(&models.User{})
-	db.AutoMigrate(&models.Group{})
-	db.AutoMigrate(&models.Transaction{})
 }
